Add symbols for quantities
Add different symbols for the quantities. For period the variable is P,
for orbital quantities the appropriate variable is R. Clarifications
are made via indices.

Also, only print a number for moons or moonlets in the overview table
if the number is not 0. This avoids a lot of meaningless entries.