Disallow two 'Distant' companion stars
In a certain edge case the program would get stuck in an infinite loop
when trying to generate the orbital parameters for the second companion
star. If the first companion is in the farthest category ('Distant')
then the orbit of the second companion needs to be larger. So, if the
first companion gets a high orbit (by chance) then it will take an
infinite amount of time for the second companion to get a valid orbit,
as there is none to be had.